feat(custom progress): make code better and cleaner
