+++ src/modules/vocabularies/components/statistics/components/add-info/addInfo.module.css
@@ -1,0 +1,66 @@
+.statistics {
+	display: flex;
+	margin-top: 30px;
+	gap: 100px;
+	width: 70%;
+	justify-content: center;
+}
+
+.modeContainer {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	gap: 20px;
+}
+
+.modeTitle {
+	font-size: 18px;
+	font-weight: bold;
+}
+
+.randomProgress {
+	display: flex;
+}
+
+.explanation {
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	margin-left: 5px;
+	gap: 5px;
+}
+
+.markersExplanation {
+	display: flex;
+	align-items: center;
+}
+
+.marker {
+	width: 16px;
+	height: 16px;
+}
+
+.success {
+	background: var(--color-success);
+}
+
+.error {
+	background: var(--color-error);
+}
+
+.primary {
+	background: var(--color-primary);
+}
+
+.moreInfoContainer {
+	width: 100%;
+	margin-top: 20px;
+}
+
+@media (max-width: 1201px) {
+	.statistics {
+		gap: 30px;
+		flex-direction: column;
+		width: 100%;
+	}
+}
