+++ src/modules/vocabularies/components/statistics/statistics.module.css
@@ -17,55 +17,6 @@
 	font-weight: bold;
 }
 
-.statistics {
-	display: flex;
-	margin-top: 30px;
-	gap: 100px;
-	width: 70%;
-	justify-content: center;
-}
-
-.modeContainer {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	gap: 20px;
-}
-
-.randomProgress {
-	display: flex;
-}
-
-.explanation {
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	margin-left: 5px;
-	gap: 5px;
-}
-
-.markersExplanation {
-	display: flex;
-	align-items: center;
-}
-
-.marker {
-	width: 16px;
-	height: 16px;
-}
-
-.success {
-	background: var(--color-success);
-}
-
-.error {
-	background: var(--color-error);
-}
-
-.primary {
-	background: var(--color-primary);
-}
-
 .modeTitle {
 	font-size: 18px;
 	font-weight: bold;
@@ -77,11 +28,6 @@
 	justify-content: center;
 	gap: 0px 15px;
 	margin-top: 10px;
-}
-
-.moreInfoContainer {
-	width: 100%;
-	margin-top: 20px;
 }
 
 .uncorrectAnswersContainer {
@@ -111,12 +57,6 @@
 	}
 
 	.totalProgress {
-		flex-direction: column;
-		width: 100%;
-	}
-
-	.statistics {
-		gap: 30px;
 		flex-direction: column;
 		width: 100%;
 	}
